refactor(module/event/fetcher): rename code file
